Python/v3-p/requirements.txt
fastapi  
httpx  
langchain  
langchain-groq  
llama-index-llms-groq
oci  
pydantic-settings  
python-multipart  
uvicorn
sendgrid
ddgs
pytest
pytest-asyncio

#PDF DocSort
pytesseract
pypdfium2
Pillow
numpy
opencv-python